Statistics: move Statistics to its own file onmt/Statistics
The Statistics.score() staticmethod is only used by LossCompute,
make it a class method of LossComput its own.
