+++ onmt/Loss.py
@@ -1,13 +1,9 @@
 """
 This file handles the details of the loss function during training.
 
-This includes: loss criterion, training statistics, and memory optimizations.
+This includes: loss criterion and memory optimizations.
 """
 from __future__ import division
-import time
-import sys
-import math
-
 import torch
 import torch.nn as nn
 from torch.autograd import Variable
@@ -25,60 +21,6 @@ def nmt_criterion(vocab_size, gpuid, pad_id):
     if gpuid:
         crit.cuda()
     return crit
-
-
-class Statistics:
-    """
-    Training loss function statistics.
-    """
-    def __init__(self, loss=0, n_words=0, n_correct=0):
-        self.loss = loss
-        self.n_words = n_words
-        self.n_correct = n_correct
-        self.n_src_words = 0
-        self.start_time = time.time()
-
-    def update(self, stat):
-        self.loss += stat.loss
-        self.n_words += stat.n_words
-        self.n_correct += stat.n_correct
-
-    def accuracy(self):
-        return 100 * (self.n_correct / self.n_words)
-
-    def ppl(self):
-        return math.exp(min(self.loss / self.n_words, 100))
-
-    def elapsed_time(self):
-        return time.time() - self.start_time
-
-    def output(self, epoch, batch, n_batches, start):
-        t = self.elapsed_time()
-        print(("Epoch %2d, %5d/%5d; acc: %6.2f; ppl: %6.2f; " +
-               "%3.0f src tok/s; %3.0f tgt tok/s; %6.0f s elapsed") %
-              (epoch, batch,  n_batches,
-               self.accuracy(),
-               self.ppl(),
-               self.n_src_words / (t + 1e-5),
-               self.n_words / (t + 1e-5),
-               time.time() - start))
-        sys.stdout.flush()
-
-    def log(self, prefix, experiment, optim):
-        t = self.elapsed_time()
-        experiment.add_scalar_value(prefix + "_ppl", self.ppl())
-        experiment.add_scalar_value(prefix + "_accuracy", self.accuracy())
-        experiment.add_scalar_value(prefix + "_tgtper",  self.n_words / t)
-        experiment.add_scalar_value(prefix + "_lr", optim.lr)
-
-    @staticmethod
-    def score(loss, scores, targ, pad):
-        pred = scores.max(1)[1]
-        non_padding = targ.ne(pad)
-        num_correct = pred.eq(targ) \
-                          .masked_select(non_padding) \
-                          .sum()
-        return Statistics(loss[0], non_padding.sum(), num_correct)
 
 
 def filter_gen_state(state):
@@ -136,7 +78,7 @@ def shards(state, shard_size, eval=False):
         torch.autograd.backward(inputs, grads)
 
 
-class LossCompute:
+class LossCompute(object):
     def __init__(self, generator, crit, tgt_vocab, dataset, epoch, copy_attn):
         self.generator = generator
         self.crit = crit
@@ -193,5 +135,14 @@ class LossCompute:
         # Coverage loss term.
         ppl = loss.data.clone()
 
-        stats = Statistics.score(ppl, scores_data, target, pad)
+        stats = self.score(ppl, scores_data, target, pad)
+
         return loss, stats
+
+    def score(self, loss, scores, targ, pad):
+        pred = scores.max(1)[1]
+        non_padding = targ.ne(pad)
+        num_correct = pred.eq(targ) \
+                          .masked_select(non_padding) \
+                          .sum()
+        return onmt.Statistics(loss[0], non_padding.sum(), num_correct)
