+++ onmt/__init__.py
@@ -5,8 +5,9 @@ from onmt.Trainer import Trainer
 from onmt.Translator import Translator
 from onmt.Optim import Optim
 from onmt.Beam import Beam, GNMTGlobalScorer
+from onmt.Statistics import Statistics
 
 
 # For flake8 compatibility
 __all__ = [onmt.Loss, onmt.IO, onmt.Models, Trainer, Translator,
-           Optim, Beam, GNMTGlobalScorer]
+           Optim, Beam, Statistics, GNMTGlobalScorer]
